refactor: refactor directory structure & change name profileBox to userBox
