+++ lib/render/render.js
