+++ lib/router/router.js
@@ -1,4 +1,4 @@
-import { renderAll } from './render.js';
+import { renderAll } from '../render/render.js';
 import { setPageIconBackground } from '../../src/utils/icon_status.js';
 import { navigate } from './navigate.js';
 import { PageNotFound } from '../../src/pages/404.js';
