+++ src/pages/tournament.js
@@ -4,7 +4,7 @@ export function pageTournament() {
   const content = `
         <div class="side-bar"></div>
         <div class="main-box">
-          <div class="profile-box"></div>
+          <div class="user-box"></div>
           <div class="game-box">
             <p class="header-pong">PONG</p>
             <p class="header-classic">TOURNAMENT</p>
